feat(user): 소비유형 결정 로직 구현

+++ src/main/java/com/goolbitg/api/entity/UserStats.java
@@ -28,6 +28,6 @@ public class UserStats {
     private Integer postCount;
 
     @Column(name = "achivement_guage")
-    private Float achivementGuage;
+    private Integer achivementGuage;
 
 }
